Move cluster into net
This commit moves the cluster module inside net to match endpoint. There
are no functional changes.
